requirements.txt
django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
tenacity==8.2.3
spacy==3.7.2
pdfplumber==0.10.2
scikit-learn==1.3.2
python-dotenv==1.0.0
psycopg2
pandas
django-cors-headers==4.3.1
textacy
rapidfuzz
